+++ wrappers/src/query_cs.cpp
@@ -26,6 +26,7 @@
 #include "timestamp_helpers.hpp"
 #include "object-store/src/results.hpp"
 #include "marshalable_sort_clause.hpp"
+#include "object_accessor.hpp"
 
 
 using namespace realm;
@@ -395,6 +396,39 @@ REALM_EXPORT Results* query_create_sorted_results(Query * query_ptr, SharedRealm
         auto sort_descriptor = SortDescriptor(*table_ptr, column_indices, ascending);
         return new Results(*realm, *query_ptr, sort_descriptor);
     });
+}
+
+
+#pragma mark  PrimaryKey Searches
+Row* row_for_primarykey(Table* table_ptr, size_t columnIndex, std::function<size_t(Table*)> finder, NativeException::Marshallable& ex)
+{
+    return handle_errors(ex, [&]() {
+        if (columnIndex >= table_ptr->get_column_count()) {
+            const std::string name {table_ptr->get_name()};
+            throw MissingPrimaryKeyException( name, name + " does not have a primary key");
+        }
+        const size_t row_ndx = finder(table_ptr);
+        if (row_ndx == not_found)
+            return (Row*)nullptr;
+        return new Row(table_ptr->get(row_ndx));
+    });
+}
+
+
+REALM_EXPORT Row* row_for_int_primarykey(Table* table_ptr, size_t columnIndex, int64_t value, NativeException::Marshallable& ex)
+{
+    return row_for_primarykey(table_ptr, columnIndex, [=](Table* table) {
+        return table->find_first_int(columnIndex, value);
+    }, ex);
+}
+  
+
+REALM_EXPORT Row* row_for_string_primarykey(Table* table_ptr, size_t columnIndex, uint16_t* value, size_t value_len, NativeException::Marshallable& ex)
+{
+    Utf16StringAccessor str(value, value_len);
+    return row_for_primarykey(table_ptr, columnIndex, [&](Table* table) {
+        return table->find_first_string(columnIndex, str);
+    }, ex);
 }
 
 }   // extern "C"